Refactors game components and updates routing
Moves game-related components into a dedicated 'games' directory for better organization.

Updates the application routing to reflect the new location of the game components.

Removes unused components and dependencies to streamline the application and improve maintainability.
