+++ src/app/app.routes.ts
@@ -14,7 +14,7 @@ import { LandingPageComponent } from './pages/landing-page/landing-page.componen
 import { ProfileComponent } from './pages/profile/profile.component';
 import { GameGalleryComponent } from './game-gallery/game-gallery.component';
 import { TeamComponent } from './pages/team/team.component';
-import { CrosswordGameComponent } from './pages/crossword-game/crossword-game.component';
+import { CrosswordGameComponent } from './pages/games/crossword-game/crossword-game.component';
 import { GameSequenceComponent } from './pages/games/game-sequence/game-sequence.component'; 
 import { MemoryGameComponent } from './pages/games/memorycard-game/memorycard-game.component';
 import { SuggestionComponent } from './pages/suggestion/suggestion.component';
@@ -40,7 +40,7 @@ export const routes: Routes = [
   {
     path: 'puzzle-board',
     loadComponent: () =>
-      import('./components/puzzle-board/puzzle-board.component').then(
+      import('./pages/games/puzzle-board/puzzle-board.component').then(
         (m) => m.PuzzleBoardComponent
       ),
   },
