+++ src/app/components/app-layout/app-layout.component.ts
@@ -4,7 +4,6 @@ import { TopbarComponent } from './elements/topbar/topbar.component';
 import { SidebarComponent } from './elements/sidebar/sidebar.component';
 import { CommonModule } from '@angular/common';
 import { LayoutService } from '../../services/layout.service';
-import { SvgIconComponent } from '../svg-icon/svg-icon.component';
 
 @Component({
   selector: 'app-layout',
@@ -13,8 +12,7 @@ import { SvgIconComponent } from '../svg-icon/svg-icon.component';
     CommonModule,
     RouterOutlet,
     TopbarComponent,
-    SidebarComponent,
-    SvgIconComponent
+    SidebarComponent
   ],
   templateUrl: './app-layout.component.html',
 })
