--- src/app/components/gameProfeCodigo/games-list/games-list.component.html
@@ -1,52 +1,0 @@
-<div class="table-responsive">
-  <table class="table">
-    <thead>
-      <tr>
-        <th scope="col">#</th>
-        <th scope="col">Name</th>
-        <th scope="col">Description</th>
-        <th scope="col">Created At</th>
-        <th scope="col">Updated At</th>
-        <th scope="col">Status</th>
-        <th scope="col">Actions</th>
-      </tr>
-    </thead>
-    <tbody>
-      @for (item of itemList; track $index) {
-        <tr>
-          <th scope="row">{{item.id}}</th>
-          <td>{{item.name}}</td>
-          <td>{{item.description}}</td>
-          <td>{{item.createdAt | date}}</td>
-          <td>{{item.updatedAt | date}}</td>
-          <td>
-            <p 
-              class="text-white text-center"
-              [ngClass]="{
-              'bg-info': item.status == 'active',
-              'bg-warning': item.status == 'disabled'
-              }">{{item.status}}</p>
-          </td>
-          <td>
-            <button type="button" (click)="showDetailModal(item, detailModal)"  class="btn btn-secondary">
-              <i class="fa-solid fa-eye white"
-              ></i>
-            </button>
-            <button type="button" (click)="deleteGame(item)" class="btn btn-danger">
-              <i class="fa-solid fa-trash white"
-              ></i>
-            </button>
-          </td>
-        </tr>
-      }
-    </tbody>
-  </table>
-</div>
-
-<app-modal #detailModal>
-  <app-games-form
-  (callParentEvent)="onFormEventCalled($event)"
-  [title]="'Update Game'"
-  [toUpdateGame]="selectedItem"
-  />
-</app-modal>
