--- src/app/components/orders/orders-form/orders-form.component.html
@@ -1,22 +1,0 @@
-<form [formGroup]="orderForm" class="row">
-  <div>
-    <h2 class="mb-0">Add Order</h2>
-  </div>
-  <div class="mb-2">
-    <label class="mb-1" for="description">Description*</label>
-    <input id="description" formControlName="description" type="text" class="form-control">
-    @if (orderForm.controls['description'].invalid && (orderForm.controls['description'].dirty || orderForm.controls['description'].touched)) {
-      <p class="text-danger my-1">Description is required</p>
-    }
-  </div>
-  <div class="mb-2">
-    <label class="mb-1" for="description">Total*</label>
-    <input min="1" id="description" formControlName="total" type="number" class="form-control">
-    @if (orderForm.controls['total'].invalid && (orderForm.controls['total'].dirty || orderForm.controls['total'].touched)) {
-      <p class="text-danger my-1">Total minimum required 1</p>
-    }
-  </div>
-  <div class="d-flex justify-content-end">
-    <button (click)="callSave()" [disabled]="orderForm.invalid" class="btn btn-primary" type="button">Save</button>
-  </div>
-</form>
