--- src/app/components/orders/orders-list/orders-list.component.html
@@ -1,34 +1,0 @@
-
-<h2>{{title}}</h2>
-
-<table class="table">
-  <thead>
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Description</th>
-      <th scope="col">Total</th>
-      <th scope="col">Actions</th>
-    </tr>
-  </thead>
-  <tbody>
-    @for (item of orders; track $index) {
-      <tr>
-        <th scope="row">{{item.id}}</th>
-        <td>{{item.description}}</td>
-        <td>{{item.total}}</td>
-        <td>
-          <button (click)="callModalAction.emit(item);" class="btn btn-warning me-2" type="button"><i class="fas fa-pen"></i></button>
-          <button (click)="callDeleteAction.emit(item)" class="btn btn-danger" type="button"><i class="fas fa-trash"></i></button>
-        </td>
-      </tr>
-    } @empty {
-      <tr>
-        <td colspan="4">
-          <div class="d-flex justify-content-center align-items-center w-100">
-            <p class="mb-0">There are not orders</p>
-          </div>
-        </td>
-      </tr>
-    }
-  </tbody>
-</table>
