--- src/app/components/orders/orders-list/orders-list.component.ts
@@ -1,16 +1,0 @@
-import { AfterViewInit, Component, Input, OnInit, OnChanges, SimpleChanges, Output, EventEmitter } from '@angular/core';
-import { IOrder } from '../../../interfaces';
-
-@Component({
-  selector: 'app-orders-list',
-  standalone: true,
-  imports: [],
-  templateUrl: './orders-list.component.html',
-  styleUrl: './orders-list.component.scss'
-})
-export class OrdersListComponent {
-  @Input() title: string  = '';
-  @Input() orders: IOrder[] = [];
-  @Output() callModalAction: EventEmitter<IOrder> = new EventEmitter<IOrder>();
-  @Output() callDeleteAction: EventEmitter<IOrder> = new EventEmitter<IOrder>();
-}
