--- src/app/components/preferenceList/preference-list/preference-list.component.html
@@ -1,31 +1,0 @@
-<div class="table-container table-responsive">
-  <table class="table table-striped">
-    <thead>
-      <tr>
-        <th scope="col">ID</th>
-        <th scope="col">Name</th>
-        <th class="text-center" scope="col">Actions</th>
-      </tr>
-    </thead>
-    <tbody>
-      @for (item of preferenceList; track item.id) {
-        <tr>
-          <th scope="row">{{item.id}}</th>
-          <td>{{item.name}}</td>
-          <td>
-            <button (click)="callModalAction.emit(item);" class="btn btn-warning me-2" type="button"><i class="fas fa-pen"></i></button>
-            <button (click)="callDeleteAction.emit(item)" class="btn btn-danger" type="button"><i class="fas fa-trash"></i></button>
-          </td>
-        </tr>
-      } @empty {
-        <tr>
-          <td colspan="7">
-            <div class="d-flex justify-content-center align-items-center w-100">
-              <p class="mb-0">There are not prefereceList</p>
-            </div>
-          </td>
-        </tr>
-      }
-    </tbody>
-  </table>
-</div>
