+++ src/app/pages/games/puzzle-board/puzzle-board.component.html
