+++ src/app/pages/games/puzzle-board/puzzle-board.component.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/utils/_variables';
+@import '../../../../assets/utils/_variables';
 
 .puzzle-main-layout {
   display: flex;
